refactor(sdk): move wallet class out of entities folder
