+++ packages/new-polymath-sdk/src/Context.ts
@@ -3,7 +3,7 @@ import { PolyToken } from '~/LowLevel/PolyToken';
 import { PolymathRegistry } from '~/LowLevel/PolymathRegistry';
 import { SecurityTokenRegistry } from '~/LowLevel/SecurityTokenRegistry';
 import { ModuleRegistry } from '~/LowLevel/ModuleRegistry';
-import { Wallet } from '~/entities/Wallet';
+import { Wallet } from '~/Wallet';
 
 interface Params {
   polyToken: PolyToken;
